+++ mail_deduplicate/tests/test_mail.py
@@ -40,7 +40,7 @@ def test_invalid_date_parsing_noop(invoke, make_box):
         ],
     )
 
-    result = invoke("--strategy=discard-newest", box_path)
+    result = invoke("--strategy=discard-newest", "--action=delete-discarded", box_path)
 
     assert result.exit_code == 0
 
@@ -77,6 +77,7 @@ def test_invalid_date_parsing_dedup(invoke, make_box):
         "--hash-header=to",
         "--hash-header=subject",
         "--strategy=discard-newest",
+        "--action=delete-discarded",
         box_path,
     )